add s3 environment variables to docker-compose.yml, in progress helpers for uploading csv to s3
